Feature/253 create shared batch form (#254)
* Extracted BatchForm in its own component

---------
