+++ canopeum_backend/canopeum_backend/models.py
@@ -178,7 +178,7 @@ class Site(models.Model):
 
 
 class Batch(models.Model):
-    site = models.ForeignKey(Site, models.CASCADE, blank=True, null=True)
+    site = models.ForeignKey(Site, models.CASCADE)
     created_at = models.DateTimeField(auto_now_add=True, null=True)
     updated_at = models.DateTimeField(auto_now=True, null=True)
     name = models.TextField(blank=True, null=True)
