+++ canopeum_frontend/src/components/analytics/batch-modal/EditBatchModal.tsx
@@ -1,66 +1,33 @@
-/* eslint-disable max-lines -- disable max-lines */
 import { Dialog, DialogActions, DialogContent, DialogTitle } from '@mui/material'
-import { useCallback, useContext, useState } from 'react'
+import { useContext, useState } from 'react'
 import { useTranslation } from 'react-i18next'
 
-import FertilizersSelector from '@components/analytics/FertilizersSelector'
-import MulchLayersSelector from '@components/analytics/MulchLayersSelector'
-import SupportSpeciesSelector from '@components/analytics/SupportSpeciesSelector'
-import TreeSpeciesSelector from '@components/analytics/TreeSpeciesSelector'
+import BatchForm from '@components/analytics/batch-modal/BatchForm'
+import { type BatchFormDto, DEFAULT_BATCH_FORM_DTO } from '@components/analytics/batch-modal/batchModal.model'
 import { SnackbarContext } from '@components/context/SnackbarContext'
 import useApiClient from '@hooks/ApiClientHook'
-import type { BatchDetail, FertilizerType, MulchLayerType, TreeType } from '@services/api'
-import { Seeds, Species } from '@services/api'
-import { floorNumberValue } from '@utils/formUtils'
+import type { BatchDetail } from '@services/api'
 
 type Props = {
   readonly batchToEdit: BatchDetail,
   readonly handleClose: (reason?: 'edit') => void,
 }
 
-type EditBatchDto = {
-  id: number,
-  name?: string,
-  sponsor?: string,
-  size?: number,
-  soilCondition?: string,
-  plantCount?: number,
-  survivedCount?: number,
-  replaceCount?: number,
-  totalNumberSeed?: number,
-  totalPropagation?: number,
-  // TODO: image?: File,
-  fertilizers: FertilizerType[],
-  mulchLayers: MulchLayerType[],
-  seeds: Seeds[],
-  species: Species[],
-  supportedSpecies: TreeType[],
-}
-
-const transformToEditBatchDto = (batchDetail: BatchDetail): EditBatchDto => ({
-  ...batchDetail,
-  seeds: batchDetail.seeds.map(batchSeed =>
-    new Seeds({ id: batchSeed.treeType.id, quantity: batchSeed.quantity })
-  ),
-  species: batchDetail.species.map(batchSpecies =>
-    new Species({ id: batchSpecies.treeType.id, quantity: batchSpecies.quantity })
-  ),
-})
-
 const BatchModal = ({ batchToEdit, handleClose }: Props) => {
   const { t } = useTranslation()
   const { getApiClient } = useApiClient()
   const { openAlertSnackbar } = useContext(SnackbarContext)
 
-  const [batch, setBatch] = useState<EditBatchDto>(transformToEditBatchDto(batchToEdit))
+  const [batch, setBatch] = useState<BatchFormDto>(DEFAULT_BATCH_FORM_DTO)
+
+  const handleBatchChange = (batchFormDto: BatchFormDto) => setBatch(batchFormDto)
 
   const handleSubmitBatch = async () => {
     const {
-      id,
       name,
       size,
       soilCondition,
-      sponsor,
+      sponsorName,
       fertilizers,
       mulchLayers,
       supportedSpecies,
@@ -74,16 +41,11 @@ const BatchModal = ({ batchToEdit, handleClose }: Props) => {
       // image,
     } = batch
 
-    // TODO:
-    // const batchImage = image
-    //   ? await assetFormatter(image)
-    //   : undefined
-
     try {
       await getApiClient().batchClient.update(
-        id,
+        batchToEdit.id,
         name,
-        sponsor,
+        sponsorName,
         size,
         soilCondition,
         plantCount,
@@ -96,6 +58,7 @@ const BatchModal = ({ batchToEdit, handleClose }: Props) => {
         seeds,
         species,
         supportedSpecies.map(specie => specie.id),
+        // image,
       )
     } catch {
       openAlertSnackbar(
@@ -108,19 +71,14 @@ const BatchModal = ({ batchToEdit, handleClose }: Props) => {
     openAlertSnackbar(
       t('analyticsSite.batch-modal.feedback.edit-success'),
     )
-    resetBatch()
+
     handleClose('edit')
   }
 
-  const onClose = () => {
-    resetBatch()
-    handleClose()
-  }
-
-  const resetBatch = () => setBatch(transformToEditBatchDto(batchToEdit))
+  const onCancel = () => handleClose()
 
   return (
-    <Dialog fullWidth maxWidth='sm' onClose={onClose} open={!!batchToEdit}>
+    <Dialog fullWidth maxWidth='sm' onClose={onCancel} open={!!batchToEdit}>
       <DialogTitle>
         <div className='fs-5 text-capitalize m-auto text-center'>
           {t('analyticsSite.batch-modal.edit-title')}
@@ -128,229 +86,16 @@ const BatchModal = ({ batchToEdit, handleClose }: Props) => {
       </DialogTitle>
 
       <DialogContent className='pb-5'>
-        <form>
-          <div className='d-flex flex-column gap-3'>
-            <div>
-              <label className='form-label' htmlFor='batch-name'>
-                {t('analyticsSite.batch-modal.name-label')}
-              </label>
-              <input
-                className='form-control'
-                id='batch-name'
-                onChange={event => setBatch(value => ({ ...value, name: event.target.value }))}
-                type='text'
-                value={batch.name}
-              />
-            </div>
-
-            <div>
-              <label className='form-label text-capitalize' htmlFor='sponsor'>
-                {t('analyticsSite.batch-modal.sponsor-label')}
-              </label>
-              <input
-                className='form-control'
-                id='sponsor'
-                onChange={event => setBatch(value => ({ ...value, sponsor: event.target.value }))}
-                type='text'
-                value={batch.sponsor}
-              />
-            </div>
-
-            <div>
-              <label className='form-label text-capitalize' htmlFor='size'>
-                {t('analyticsSite.batch-modal.size-label')}
-              </label>
-              <div className='input-group'>
-                <input
-                  className='form-control'
-                  id='size'
-                  min={0}
-                  onChange={event =>
-                    setBatch(value => ({
-                      ...value,
-                      size: Number.parseInt(event.target.value, 10),
-                    }))}
-                  type='number'
-                  value={floorNumberValue(batch.size)}
-                />
-                <span className='input-group-text'>
-                  {t('analyticsSite.batch-modal.feet-squared')}
-                </span>
-              </div>
-            </div>
-
-            <div>
-              <TreeSpeciesSelector
-                label='analyticsSite.batch-modal.number-of-trees-label'
-                onChange={useCallback(
-                  species =>
-                    setBatch(current => ({
-                      ...current,
-                      species,
-                    })),
-                  [],
-                )}
-                species={batch.species}
-              />
-            </div>
-
-            <div>
-              <label className='form-label' htmlFor='soil-condition'>
-                {t('analyticsSite.batch-modal.soil-condition-label')}
-              </label>
-              <input
-                className='form-control'
-                id='soil-condition'
-                onChange={event =>
-                  setBatch(value => ({ ...value, soilCondition: event.target.value }))}
-                type='text'
-                value={batch.soilCondition}
-              />
-            </div>
-
-            <FertilizersSelector
-              fertilizers={batch.fertilizers}
-              onChange={useCallback(
-                fertilizers =>
-                  setBatch(current => ({
-                    ...current,
-                    fertilizers,
-                  })),
-                [],
-              )}
-            />
-
-            <MulchLayersSelector
-              mulchLayers={batch.mulchLayers}
-              onChange={useCallback(
-                mulchLayers =>
-                  setBatch(current => ({
-                    ...current,
-                    mulchLayers,
-                  })),
-                [],
-              )}
-            />
-
-            <SupportSpeciesSelector
-              onChange={useCallback(
-                supportedSpecies =>
-                  setBatch(current => ({
-                    ...current,
-                    supportedSpecies,
-                  })),
-                [],
-              )}
-              species={batch.supportedSpecies}
-            />
-
-            <div>
-              <label className='form-label text-capitalize' htmlFor='total-number-of-plants'>
-                {t('analyticsSite.batch-modal.total-number-of-plants-label')}
-              </label>
-              <input
-                className='form-control'
-                id='total-number-of-plants'
-                onChange={event =>
-                  setBatch(value => ({
-                    ...value,
-                    plantCount: Number.parseInt(event.target.value, 10),
-                  }))}
-                type='number'
-                value={floorNumberValue(batch.plantCount)}
-              />
-            </div>
-
-            <div>
-              <label className='form-label text-capitalize' htmlFor='survived'>
-                {t('analyticsSite.batch-modal.survived-label')}
-              </label>
-              <input
-                className='form-control'
-                id='survived'
-                onChange={event =>
-                  setBatch(value => ({
-                    ...value,
-                    survivedCount: Number.parseInt(event.target.value, 10),
-                  }))}
-                type='number'
-                value={floorNumberValue(batch.survivedCount)}
-              />
-            </div>
-
-            <div>
-              <label className='form-label text-capitalize' htmlFor='replaced'>
-                {t('analyticsSite.batch-modal.replaced-label')}
-              </label>
-              <input
-                className='form-control'
-                id='replaced'
-                onChange={event =>
-                  setBatch(value => ({
-                    ...value,
-                    replaceCount: Number.parseInt(event.target.value, 10),
-                  }))}
-                type='number'
-                value={floorNumberValue(batch.replaceCount)}
-              />
-            </div>
-
-            <div>
-              <TreeSpeciesSelector
-                label='analyticsSite.batch-modal.seeds-per-species-label'
-                onChange={useCallback(
-                  species =>
-                    setBatch(current => ({
-                      ...current,
-                      seeds: species,
-                    })),
-                  [],
-                )}
-                species={batch.seeds}
-              />
-            </div>
-
-            <div>
-              <label className='form-label text-capitalize' htmlFor='totalNumberSeed'>
-                {t('analyticsSite.batch-modal.total-seeds-label')}
-              </label>
-              <input
-                className='form-control'
-                id='totalNumberSeed'
-                onChange={event =>
-                  setBatch(value => ({
-                    ...value,
-                    totalNumberSeed: Number.parseInt(event.target.value, 10),
-                  }))}
-                type='number'
-                value={floorNumberValue(batch.totalNumberSeed)}
-              />
-            </div>
-
-            <div>
-              <label className='form-label text-capitalize' htmlFor='propagation'>
-                {t('analyticsSite.batch-modal.propagation-label')}
-              </label>
-              <input
-                className='form-control'
-                id='propagation'
-                onChange={event =>
-                  setBatch(value => ({
-                    ...value,
-                    totalPropagation: Number.parseInt(event.target.value, 10),
-                  }))}
-                type='number'
-                value={floorNumberValue(batch.totalPropagation)}
-              />
-            </div>
-          </div>
-        </form>
+        <BatchForm
+          handleBatchChange={handleBatchChange}
+          initialBatch={batchToEdit}
+        />
       </DialogContent>
 
       <DialogActions>
         <button
           className='btn btn-outline-primary'
-          onClick={() => onClose()}
+          onClick={() => onCancel()}
           type='button'
         >
           {t('generic.cancel')}
